Add useMetabotReactions hook
